refactor(cli): split skill picker view, strip restating comments
skill_picker.go landed at 1161 lines, over the 800 hard ceiling. Move
the rendering half into skill_picker_view.go (690 + 437) and drop the
what-restating doc/banner/example comments the merged code carried, so
the new skills surface matches the repo comment policy. Also drops a
dead p.confirm write in deleteSkillPick.
